requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
pydantic==2.8.2
pandas==2.2.2
openpyxl==3.1.5
python-dateutil==2.9.0.post0